move PRT test cases to separate file
